portfolio/api/requirements.txt
fastapi>=0.109.2,<0.115
uvicorn>=0.27.1
requests>=2.31.0
python-dotenv>=1.0.1
slowapi>=0.1.9
